Replace backticks with **
In `embedded\templates.autogen.go` the content of the files in
`.\templates\*` are stored in an array of strings which are build using
backticks.

So the first (unquoted) backtick will end the string element in
templates.autogen.go.

Since no other file than README.md currently has backticks in it I
just repaced them with ** (bold in markdown).

Other solution would be the quote all incomming backticks when
generating templates.autogen.go.